requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
aiosqlite>=0.19.0
yt-dlp>=2024.3.10
PyNaCl>=1.5.0
